backend/requirements.txt
fastapi
uvicorn

# For database
sqlalchemy
psycopg2-binary

# for env config
python-dotenv